git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/wxQT@77420 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/propgrid/property.cpp
@@ -755,6 +755,8 @@ void wxPGProperty::GetDisplayInfo( unsigned int column,
                                    wxString* pString,
                                    const wxPGCell** pCell )
 {
+    wxCHECK_RET( GetGrid(),
+                 wxT("Cannot obtain display info for detached property") );
     const wxPGCell* cell = NULL;
 
     if ( !(flags & wxPGCellRenderer::ChoicePopup) )
@@ -989,8 +991,13 @@ wxString wxPGProperty::GetValueAsString( int argFlags ) const
         return g_invalidStringContent;
     }
 #endif
-
     wxPropertyGrid* pg = GetGrid();
+    wxASSERT_MSG( pg,
+                  wxT("Cannot get valid value for detached property") );
+    if ( !pg )
+    {
+        return wxEmptyString;
+    }
 
     if ( IsValueUnspecified() )
         return pg->GetUnspecifiedValueText(argFlags);
@@ -1464,6 +1471,8 @@ void wxPGProperty::SetValue( wxVariant value, wxVariant* pList, int flags )
 
 void wxPGProperty::SetValueInEvent( wxVariant value ) const
 {
+    wxCHECK_RET( GetGrid(),
+                 wxT("Cannot store pending value for detached property"));
     GetGrid()->ValueChangeInEvent(value);
 }
 
@@ -1639,6 +1648,8 @@ const wxPGCell& wxPGProperty::GetCell( unsigned int column ) const
         return m_cells[column];
 
     wxPropertyGrid* pg = GetGrid();
+    wxASSERT_MSG( pg,
+                  wxT("Cannot get cell for detached property") );
 
     if ( IsCategory() )
         return pg->GetCategoryDefaultCell();
@@ -1885,22 +1896,24 @@ wxValidator* wxPGProperty::DoGetValidator() const
 int wxPGProperty::InsertChoice( const wxString& label, int index, int value )
 {
     wxPropertyGrid* pg = GetGrid();
-    int sel = GetChoiceSelection();
+    const int sel = GetChoiceSelection();
 
-    int newSel = sel;
+    int newSel = (sel == wxNOT_FOUND) ? 0 : sel;
 
+    const int numChoices = m_choices.GetCount();
     if ( index == wxNOT_FOUND )
-        index = m_choices.GetCount();
+        index = numChoices;
 
-    if ( index <= sel )
+    if ( numChoices > 0 && index <= sel )
         newSel++;
 
     m_choices.Insert(label, index, value);
-
-    if ( sel != newSel )
+    // Set new selection if it was modified
+    // or if the first element was added.
+    if ( sel != newSel || numChoices == 0 )
         SetChoiceSelection(newSel);
 
-    if ( this == pg->GetSelection() )
+    if ( pg && this == pg->GetSelection() )
         GetEditorClass()->InsertItem(pg->GetEditorControl(),label,index);
 
     return index;
@@ -1930,7 +1943,7 @@ void wxPGProperty::DeleteChoice( int index )
     if ( sel != newSel )
         SetChoiceSelection(newSel);
 
-    if ( this == pg->GetSelection() )
+    if ( pg && this == pg->GetSelection() )
         GetEditorClass()->DeleteItem(pg->GetEditorControl(), index);
 }
 
@@ -1945,7 +1958,7 @@ int wxPGProperty::GetChoiceSelection() const
 
     if ( valueType == wxPG_VARIANT_TYPE_LONG )
     {
-        index = value.GetLong();
+        index = m_choices.Index(value.GetLong());
     }
     else if ( valueType == wxPG_VARIANT_TYPE_STRING )
     {
@@ -1973,7 +1986,7 @@ void wxPGProperty::SetChoiceSelection( int newValue )
     }
     else  // if ( valueType == wxPG_VARIANT_TYPE_LONG )
     {
-        SetValue( (long) newValue );
+        SetValue( m_choices.GetValue(newValue) );
     }
 }
 
@@ -2086,6 +2099,10 @@ bool wxPGProperty::RecreateEditor()
 
 void wxPGProperty::SetValueImage( wxBitmap& bmp )
 {
+    // We need PG to obtain default image size
+    wxCHECK_RET( GetGrid(),
+                 wxT("Cannot set image for detached property") );
+
     delete m_valueBitmap;
 
     if ( &bmp && bmp.IsOk() )
@@ -2209,7 +2226,10 @@ int wxPGProperty::GetY2( int lh ) const
 
 int wxPGProperty::GetY() const
 {
-    return GetY2(GetGrid()->GetRowHeight());
+    wxPropertyGrid *pg = GetGrid();
+    wxASSERT_MSG( pg,
+        wxT("Cannot obtain coordinates of detached property") );
+    return pg ? GetY2(pg->GetRowHeight()) : 0;
 }
 
 // This is used by Insert etc.
@@ -2545,8 +2565,16 @@ void wxPGProperty::Empty()
 
 wxPGProperty* wxPGProperty::GetItemAtY( unsigned int y ) const
 {
+    wxPropertyGrid *pg = GetGrid();
+    wxASSERT_MSG( pg,
+                  wxT("Cannot obtain property item for detached property") );
+    if( !pg )
+    {
+        return NULL;
+    }
+
     unsigned int nextItem = 0;
-    return GetItemAtY( y, GetGrid()->GetRowHeight(), &nextItem);
+    return GetItemAtY(y, pg->GetRowHeight(), &nextItem);
 }
 
 void wxPGProperty::DeleteChildren()
